refactor : change di layer package
