+++ app/src/main/java/woowacourse/shopping/ui/MainActivity.kt
@@ -6,7 +6,7 @@ import android.view.Menu
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.shopping.R
-import woowacourse.shopping.data.di.component.viewModelsWithAutoInject
+import woowacourse.shopping.di.component.viewModelsWithAutoInject
 import woowacourse.shopping.databinding.ActivityMainBinding
 import woowacourse.shopping.ui.cart.CartActivity
