+++ app/src/main/java/woowacourse/shopping/ui/cart/CartActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.shopping.R
-import woowacourse.shopping.data.di.component.viewModelsWithAutoInject
+import woowacourse.shopping.di.component.viewModelsWithAutoInject
 import woowacourse.shopping.databinding.ActivityCartBinding
 
 class CartActivity : AppCompatActivity() {
